requirements.txt
aiogram>=3.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
remnawave>=2.6.0
